+++ source/Public/Publish-TkMemPolicyManagerApp.ps1
@@ -241,7 +241,7 @@ function Publish-TkMemPolicyManagerApp {
                 CertThumbprint             = $CertDetails.CertThumbprint
                 ErrorAction                = 'Stop'
             }
-            $ConsentUrl = Initialize-TkAppSpRegistration @AppSpRegistrationParams
+            $ConsentUrl = New-TkAppSpOauth2Registration @AppSpRegistrationParams
             [void](Read-Host 'Provide admin consent now, or copy the url and provide admin consent later. Press Enter to continue.')
             # 8) Build a final PSCustomObject to store in the secret vault
             $TkMemPolicyManagerAppParams = @{